+++ app/src/main/java/org/stepik/android/view/catalog_block/ui/adapter/delegate/AuthorListAdapterDelegate.kt
@@ -1,63 +1,77 @@
 package org.stepik.android.view.catalog_block.ui.adapter.delegate
 
-import android.graphics.BitmapFactory
 import android.view.View
 import android.view.ViewGroup
-import androidx.core.graphics.drawable.RoundedBitmapDrawableFactory
-import com.bumptech.glide.Glide
+import androidx.recyclerview.widget.RecyclerView
 import kotlinx.android.extensions.LayoutContainer
+import kotlinx.android.synthetic.main.header_catalog_block.*
 import kotlinx.android.synthetic.main.item_author_list.*
-import kotlinx.android.synthetic.main.layout_author_properties.view.*
 import org.stepic.droid.R
-import org.stepic.droid.ui.util.RoundedBitmapImageViewTarget
+import org.stepic.droid.ui.util.CoursesSnapHelper
 import org.stepik.android.domain.catalog_block.model.CatalogAuthor
+import org.stepik.android.presentation.course_list_redux.model.CatalogBlockStateWrapper
+import org.stepik.android.view.base.ui.adapter.layoutmanager.TableLayoutManager
+import org.stepik.android.view.catalog_block.mapper.AuthorCountMapper
+import org.stepik.android.view.catalog_block.model.CatalogItem
+import org.stepik.android.view.catalog_block.ui.delegate.CatalogBlockHeaderDelegate
+import ru.nobird.android.core.model.cast
 import ru.nobird.android.ui.adapterdelegates.AdapterDelegate
 import ru.nobird.android.ui.adapterdelegates.DelegateViewHolder
+import ru.nobird.android.ui.adapters.DefaultDelegateAdapter
 
 class AuthorListAdapterDelegate(
-    private val onItemClick: (Long) -> Unit
-) : AdapterDelegate<CatalogAuthor, DelegateViewHolder<CatalogAuthor>>() {
-    override fun isForViewType(position: Int, data: CatalogAuthor): Boolean =
-        true
+    private val authorCountMapper: AuthorCountMapper,
+    private val onAuthorClick: (Long) -> Unit
+) : AdapterDelegate<CatalogItem, DelegateViewHolder<CatalogItem>>() {
+    private val sharedViewPool = RecyclerView.RecycledViewPool()
 
-    override fun onCreateViewHolder(parent: ViewGroup): DelegateViewHolder<CatalogAuthor> =
-        ViewHolder(createView(parent, R.layout.item_author_list))
+    override fun isForViewType(position: Int, data: CatalogItem): Boolean =
+        data is CatalogItem.Block && data.catalogBlockStateWrapper is CatalogBlockStateWrapper.AuthorList
 
-    private inner class ViewHolder(
+    override fun onCreateViewHolder(parent: ViewGroup): DelegateViewHolder<CatalogItem> =
+        AuthorListViewHolder(createView(parent, R.layout.item_author_list))
+
+    private inner class AuthorListViewHolder(
         override val containerView: View
-    ) : DelegateViewHolder<CatalogAuthor>(containerView), LayoutContainer {
-        private val authorCoverImageTarget =
-            RoundedBitmapImageViewTarget(context.resources.getDimension(R.dimen.corner_radius), authorListImage)
+    ) : DelegateViewHolder<CatalogItem>(containerView), LayoutContainer {
 
-        private val authorPlaceholder = BitmapFactory
-            .decodeResource(context.resources, R.drawable.general_placeholder)
-            .let { bitmap ->
-                RoundedBitmapDrawableFactory
-                    .create(context.resources, bitmap)
-                    .apply {
-                        cornerRadius = context.resources.getDimension(R.dimen.corner_radius)
-                    }
+        private val catalogBlockTitleDelegate =
+            CatalogBlockHeaderDelegate(catalogBlockContainer, null)
+
+        private val adapter = DefaultDelegateAdapter<CatalogAuthor>()
+            .also {
+                it += AuthorAdapterDelegate(onAuthorClick)
             }
 
-        private val authorCourseCount = authorListPropertiesContainer.coursesCountText
-        private val authorSubscriberCount = authorListPropertiesContainer.subscribersCountText
-
         init {
-            containerView.setOnClickListener { itemData?.id?.let(onItemClick) }
+            val rowCount = context.resources.getInteger(R.integer.author_lists_default_rows)
+            authorListRecycler.layoutManager =
+                TableLayoutManager(
+                    context,
+                    horizontalSpanCount = context.resources.getInteger(R.integer.author_lists_default_columns),
+                    verticalSpanCount = rowCount,
+                    orientation = RecyclerView.HORIZONTAL,
+                    reverseLayout = false
+                )
+            authorListRecycler.setRecycledViewPool(sharedViewPool)
+            authorListRecycler.setHasFixedSize(true)
+            authorListRecycler.adapter = adapter
+
+            val snapHelper = CoursesSnapHelper(rowCount)
+            snapHelper.attachToRecyclerView(authorListRecycler)
         }
 
-        override fun onBind(data: CatalogAuthor) {
-            Glide
-                .with(context)
-                .asBitmap()
-                .load(data.avatar)
-                .placeholder(authorPlaceholder)
-                .fitCenter()
-                .into(authorCoverImageTarget)
+        override fun onBind(data: CatalogItem) {
+            val authorLists = data
+                .cast<CatalogItem.Block>()
+                .catalogBlockStateWrapper
+                .cast<CatalogBlockStateWrapper.AuthorList>()
 
-            authorListTitle.text = data.fullName
-            authorCourseCount.text = context.resources.getQuantityString(R.plurals.course_count, data.createdCoursesCount, data.createdCoursesCount)
-            authorSubscriberCount.text = context.resources.getQuantityString(R.plurals.catalog_author_subscribers, data.followersCount, data.followersCount)
+            adapter.items = authorLists.content.authors
+            catalogBlockTitleDelegate.setInformation(authorLists.catalogBlockItem)
+
+            val count = authorLists.content.authors.size
+            catalogBlockTitleDelegate.setCount(authorCountMapper.mapAuthorCountToString(context, count))
         }
     }
 }
