+++ app/src/main/res/layout-w360dp/layout_author_properties.xml
@@ -1,6 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="wrap_content"
+<androidx.constraintlayout.widget.ConstraintLayout
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    android:layout_width="match_parent"
     android:layout_height="wrap_content"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools">
@@ -10,16 +11,23 @@
         android:layout_height="11dp"
         android:layout_width="14dp"
         android:layout_marginTop="5dp"
-        android:layout_alignParentLeft="true"
-        android:layout_alignParentStart="true"
         android:contentDescription="@string/learners_count"
+
+        app:layout_constraintLeft_toLeftOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintRight_toLeftOf="@id/coursesCountText"
+        app:layout_constraintEnd_toStartOf="@id/coursesCountText"
+        app:layout_constraintTop_toTopOf="parent"
         app:srcCompat="@drawable/ic_authors_info_course_count"/>
 
     <androidx.appcompat.widget.AppCompatTextView
         android:id="@+id/coursesCountText"
         style="@style/CourseWidgetPropertyTextView"
-        android:layout_toEndOf="@id/coursesCountImage"
-        android:layout_toRightOf="@id/coursesCountImage"
+        app:layout_constraintLeft_toRightOf="@id/coursesCountImage"
+        app:layout_constraintStart_toEndOf="@id/coursesCountImage"
+        app:layout_constraintRight_toLeftOf="@id/subscribersCountImage"
+        app:layout_constraintEnd_toStartOf="@id/subscribersCountImage"
+        app:layout_constraintTop_toTopOf="parent"
         tools:text="999K"/>
 
     <androidx.appcompat.widget.AppCompatImageView
@@ -27,19 +35,20 @@
         style="@style/CourseWidgetPropertyImageView"
         android:layout_height="10dp"
         android:layout_width="14dp"
-        android:layout_marginTop="13dp"
-        android:layout_below="@id/coursesCountImage"
-        android:layout_alignParentLeft="true"
-        android:layout_alignParentStart="true"
+        android:layout_marginTop="5dp"
+        app:layout_constraintLeft_toRightOf="@id/coursesCountText"
+        app:layout_constraintStart_toEndOf="@id/coursesCountText"
+        app:layout_constraintRight_toLeftOf="@id/subscribersCountText"
+        app:layout_constraintEnd_toStartOf="@id/subscribersCountText"
+        app:layout_constraintTop_toTopOf="parent"
         android:contentDescription="@string/learners_count"
         app:srcCompat="@drawable/ic_course_info_learners_count"/>
 
     <androidx.appcompat.widget.AppCompatTextView
         android:id="@+id/subscribersCountText"
         style="@style/CourseWidgetPropertyTextView"
-        android:layout_marginTop="4dp"
-        android:layout_below="@id/coursesCountText"
-        android:layout_toEndOf="@id/subscribersCountImage"
-        android:layout_toRightOf="@id/subscribersCountImage"
+        app:layout_constraintLeft_toRightOf="@id/subscribersCountImage"
+        app:layout_constraintStart_toEndOf="@id/subscribersCountImage"
+        app:layout_constraintTop_toTopOf="parent"
         tools:text="999K"/>
-</RelativeLayout>
+</androidx.constraintlayout.widget.ConstraintLayout>
