--- app/src/main/res/layout/item_author_lists.xml
@@ -1,22 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.appcompat.widget.LinearLayoutCompat
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    android:orientation="vertical">
-
-    <include layout="@layout/header_catalog_block" />
-
-    <androidx.recyclerview.widget.RecyclerView
-        android:id="@+id/authorListsRecycler"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:clipToPadding="false"
-        android:paddingTop="8dp"
-        android:paddingLeft="@dimen/course_list_padding"
-        android:paddingRight="@dimen/course_list_padding"
-        android:paddingBottom="8dp" />
-
-</androidx.appcompat.widget.LinearLayoutCompat>
